+++ tiger-common/src/main/java/de/gematik/test/tiger/common/exceptions/TigerUnknownProtocolException.java
@@ -1,0 +1,8 @@
+package de.gematik.test.tiger.common.exceptions;
+
+public class TigerUnknownProtocolException extends RuntimeException {
+
+    public TigerUnknownProtocolException(String message) {
+        super(message);
+    }
+}